# Improve performance of ProtobufHttpMessageConverter