@pivotal-cla 2022-11-16
@koo-taejin Thank you for signing the [Contributor License Agreement](https://cla.pivotal.io/sign/spring?repositoryId=spring-projects/spring-framework&pullRequestId=29496)!