@bclozel 2023-10-20
Thanks for this contribution, and sorry for getting back to you so long after.
I've introduced a JMH benchmark in 31a62ff8ba for the converter (both for reading and writing operations). Maybe the runtime profile changed since your PR, but I'm not seeing improvements with your change (it might have made things a bit worse actually).

Running with `java -jar spring-web/build/libs/spring-web-6.1.0-SNAPSHOT-jmh.jar -t 30 -f 2 -prof gc ProtobufHttpMessageConverterBenchmark`

Without this change:
```
Benchmark                                                                (messageCount)   Mode  Cnt       Score     Error   Units
ProtobufHttpMessageConverterBenchmark.readMessages                                   40  thrpt   10   66492.262 ± 397.655   ops/s
ProtobufHttpMessageConverterBenchmark.readMessages:·gc.alloc.rate                    40  thrpt   10   12079.125 ±  72.156  MB/sec
ProtobufHttpMessageConverterBenchmark.readMessages:·gc.alloc.rate.norm               40  thrpt   10  190556.022 ± 758.571    B/op
ProtobufHttpMessageConverterBenchmark.readMessages:·gc.count                         40  thrpt   10    1645.000            counts
ProtobufHttpMessageConverterBenchmark.readMessages:·gc.time                          40  thrpt   10    1412.000                ms
ProtobufHttpMessageConverterBenchmark.writeMessages                                  40  thrpt   10   48658.466 ± 715.484   ops/s
ProtobufHttpMessageConverterBenchmark.writeMessages:·gc.alloc.rate                   40  thrpt   10   12076.030 ± 195.945  MB/sec
ProtobufHttpMessageConverterBenchmark.writeMessages:·gc.alloc.rate.norm              40  thrpt   10  260480.029 ±   0.001    B/op
ProtobufHttpMessageConverterBenchmark.writeMessages:·gc.count                        40  thrpt   10    1791.000            counts
ProtobufHttpMessageConverterBenchmark.writeMessages:·gc.time                         40  thrpt   10    1427.000                ms
```

With this change:

```
Benchmark                                                                (messageCount)   Mode  Cnt       Score      Error   Units
ProtobufHttpMessageConverterBenchmark.readMessages                                   40  thrpt   10   64359.735 ± 1712.806   ops/s
ProtobufHttpMessageConverterBenchmark.readMessages:·gc.alloc.rate                    40  thrpt   10   11660.432 ±  306.253  MB/sec
ProtobufHttpMessageConverterBenchmark.readMessages:·gc.alloc.rate.norm               40  thrpt   10  190072.022 ±   12.749    B/op
ProtobufHttpMessageConverterBenchmark.readMessages:·gc.count                         40  thrpt   10    1996.000             counts
ProtobufHttpMessageConverterBenchmark.readMessages:·gc.time                          40  thrpt   10    1664.000                 ms
ProtobufHttpMessageConverterBenchmark.writeMessages                                  40  thrpt   10   48545.755 ±  362.536   ops/s
ProtobufHttpMessageConverterBenchmark.writeMessages:·gc.alloc.rate                   40  thrpt   10   12052.676 ±   90.700  MB/sec
ProtobufHttpMessageConverterBenchmark.writeMessages:·gc.alloc.rate.norm              40  thrpt   10  260512.030 ±    0.001    B/op
ProtobufHttpMessageConverterBenchmark.writeMessages:·gc.count                        40  thrpt   10    2296.000             counts
ProtobufHttpMessageConverterBenchmark.writeMessages:·gc.time                         40  thrpt   10    1743.000                 ms
```

In light of that, I'm declining this PR for now. We can always consider other changes if they're backed by benchmark results.
Thanks!